Caching resolved service instances in property wrapper.
The @Injected property wrapper now retrieves the service from the container only on first access and stores a reference to the result to speed up things. It also is testable now. This required the introduction of a reference to a service container which itself is now defined by a mockable protocol.